% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{GetBarFromWeb}
\alias{GetBarFromWeb}
\title{Get Bar History}
\arguments{
\item{symbol}{a character. Symbol Name.}

\item{barsType.}{a character. Bars Type. One from c("Ask", "Bid").}

\item{periodicity.}{a character. Periodicity. From c("S1", "S10", "M1", "M5", "M15", "M30", "H1", "H4", "D1", "W1","MN1")}

\item{startTimeMs.}{Long numeric. Timestamp from 1970-01-01 in ms.}

\item{count.}{Integer. Count of returned Bars from startTimeMs. Max is 1000. Can be negative.}
}
\value{
data.table with Bar Info
}
\description{
Get Bar History
}
