% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{GetBarsHistory}
\alias{GetBarsHistory}
\title{Get Bar History}
\arguments{
\item{symbol}{a character. Symbol Name.}

\item{barsType.}{a character. Bars Type. One from c("Ask", "Bid").}

\item{periodicity.}{a character. Periodicity. From c("S1", "S10", "M1", "M5", "M15", "M30", "H1", "H4", "D1", "W1","MN1")}

\item{startTime}{a POSIXct obj. Start Time in UTC.}

\item{endTime}{a POSIXct obj. End Time in UTC.}

\item{count.}{Integer. Count of returned Bars from startTime. Max is 1000. Can be negative. If count == 0, use time interval between startTime and endTime.}
}
\value{
data.table with Bar Info
}
\description{
Get Bar History
}
