% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{GetPipsValue}
\alias{GetPipsValue}
\title{Get Pips Value}
\arguments{
\item{targetCurrency}{a character. Currency Name.}

\item{symbols}{a character vectors. Symbols vector.}
}
\value{
a data.table with pips value in targetCurrency for every symbol in symbols vector
}
\description{
Get Pips Value
}
