% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{GetTicksFromWeb}
\alias{GetTicksFromWeb}
\title{Get Ticks History}
\arguments{
\item{startTimeMs.}{Long numeric. Timestamp from 1970-01-01 in ms.}

\item{symbol.}{A character. Symbol Name.}

\item{startTime}{a POSIXct obj. Start Time in UTC.}

\item{endTime}{a POSIXct obj. End Time in UTC.}

\item{count.}{Integer. Count of returned Ticks from startTime. Max is 1000. Can be negative. If count == 0, use time interval between startTime and endTime.}
}
\value{
data.table with Ticks Info.

data.table with Ticks Info.
}
\description{
Get Ticks History

Get Ticks History
}
