% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{GetTicksRawMethod}
\alias{GetTicksRawMethod}
\title{Get Ticks History}
\arguments{
\item{symbol.}{A character. Symbol Name.}

\item{startTimeMs.}{Long numeric. Timestamp from 1970-01-01 in ms.}

\item{count.}{Integer. Count of returned Bars from startTimeMs. Max is 1000. Can be negative.}
}
\value{
data.table with Ticks Info.
}
\description{
Get Ticks History
}
