% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{InitPrivateWebClient}
\alias{InitPrivateWebClient}
\title{Init Private Web Client Obj}
\usage{
InitPrivateWebClient(server = "ttlivewebapi.fxopen.com", port = 8443L,
  id = "", key = "", secret = "")
}
\arguments{
\item{server}{a character. Web Address.}

\item{port}{an integer. Port Number. Default is 8443}

\item{id}{a character. HMAC client id.}

\item{key}{a character. HMAC client key.}

\item{secret}{a character. HMAC secret key.}
}
\value{
rTTWebClient obj.
}
\description{
Init Private Web Client Obj
}
