% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{InitPublicWebClient}
\alias{InitPublicWebClient}
\title{Init Public Web Client Obj}
\usage{
InitPublicWebClient(server = "ttlivewebapi.fxopen.com", port = 8443L)
}
\arguments{
\item{server}{a character. Web Address.}

\item{port}{an integer. Port Number. Default is 8443}
}
\value{
rTTWebClient obj.
}
\description{
Init Public Web Client Obj
}
