% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\docType{class}
\name{RTTWebApiHost}
\alias{RTTWebApiHost}
\title{RTTWebApiHost}
\description{
RTTWebApiHost
}
\section{Fields}{

\describe{
\item{\code{client.}}{RTTWebClient obj.}
}}

\section{Methods}{

\describe{
\item{\code{GetBarsHistory(
  symbol,
  barsType = "Bid",
  periodicity = "M1",
  startTime,
  endTime = as.POSIXct(Sys.Date(), tz = "GMT"),
  count = 0L
)}}{Get Bar History}

\item{\code{GetCurrencyInfo()}}{Get All Currencies}

\item{\code{GetCurrentQuotes()}}{Get All Current Quotes}

\item{\code{GetDividends()}}{Get All Dividends}

\item{\code{GetPipsValue(targetCurrency, symbols)}}{Get Pips Value}

\item{\code{GetSymbolsInfo()}}{Get All Symbols}

\item{\code{GetTickHistory(
  symbol,
  startTime,
  endTime = as.POSIXct(Sys.Date(), tz = "GMT"),
  count = 0L
)}}{Get Bar History}
}}

