% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\docType{class}
\name{RTTWebClient}
\alias{RTTWebClient}
\title{RTTWebClient Class}
\description{
RTTWebClient Class
}
\section{Fields}{

\describe{
\item{\code{web_api_address.}}{Server address. Character}

\item{\code{web_api_port.}}{Port. Integer}

\item{\code{web_api_id.}}{Web Api Id. Character}

\item{\code{web_api_key.}}{Web Api Key. Character}

\item{\code{web_api_secrer.}}{Web Api Secret. Character}
}}

\section{Methods}{

\describe{
\item{\code{GetBarFromWeb(symbol, barsType, periodicity, startTimeMs, count)}}{Get Bar History}

\item{\code{GetCurrentQuotesFromWeb()}}{Get All Current Quotes}

\item{\code{GetDividendsFromWeb()}}{Get All Dividend}

\item{\code{GetSymbolsInfoFromWeb()}}{Get All Symbols}

\item{\code{GetTicksFromWeb(symbol, startTimeMs, count)}}{Get Ticks History}
}}

