% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTTWebClient.R
\name{RWebApiClient}
\alias{RWebApiClient}
\title{Set the server from which the info will be read}
\usage{
RWebApiClient(web_api_address, web_api_port = 8443, web_api_id = NULL,
  web_api_key = NULL, web_api_secret = NULL)
}
\arguments{
\item{web_api_address}{a character. WebApi Address.}

\item{web_api_port}{an integer. WebAPI port. Deafault is 8443.}

\item{web_api_id}{a character. WebApi Id.}

\item{web_api_key}{a character. WebApi Key.}

\item{web_api_secret}{a character. WebApi Secret.}
}
\description{
Set the server from which the info will be read
}
